feat: hamburger menu + right navbar
