+++ src/App.tsx
@@ -6,7 +6,6 @@ import Home from './pages/Home/home';
 import Hackathons from './pages/Hackathons/hackathons';
 import SponsorUs from './pages/SponsorUs/sponsors-us';
 import ContactUs from './pages/Contact-Us/contact-us';
-import TestPage from './pages/TestPage/TestPage';
 
 import './App.scss';
 
@@ -18,7 +17,6 @@ const App: FC = (): JSX.Element => {
                 <Route exact path="/sponsors" component={SponsorUs} />
                 <Route exact path="/contact-us" component={ContactUs} />
                 <Route exact path="/past-hackathons" component={Hackathons} />
-                <Route exact path="/test" component={TestPage} />
                 <Redirect from="/HackMerced" to="/" />
             </Switch>
         </Router>
