--- src/components/HamburgerMenu/colors.ts
@@ -1,6 +1,0 @@
-export const colors = {
-    yellowmellow: '#fbe69b',
-    lightbrown: '#be8b7b',
-    pearl: '#fdf2e9',
-    lightblue: '#93CEF0',
-};
